+++ src/main/scala/edu/ie3/mobsim/MobilitySimulator.scala
@@ -54,7 +54,7 @@ final class MobilitySimulator(
       ProbabilityDensityFunction[PointOfInterest]
     ],
     startTime: ZonedDateTime,
-    var electricVehicles: Seq[ElectricVehicle],
+    private var electricVehicles: Seq[ElectricVehicle],
     chargingHubTownIsPresent: Boolean,
     chargingHubHighwayIsPresent: Boolean,
     ioUtils: IoUtils,
@@ -88,28 +88,25 @@ final class MobilitySimulator(
       .toMap
 
     /* Send EV movements to SIMONA and receive charged EVs that ended parking */
-    val (
-      departedEvsFromSimona,
-      chargingStationOccupancy
-    ) = exchangeEvMovementsWithSimona(
+    val chargingStationOccupancy = exchangeEvMovementsWithSimona(
       currentTime,
       availableChargingPoints,
       currentPricesAtChargingStations,
       maxDistanceFromPoi
     )
 
-    /* Update and simulate all EVs that were returned from SIMONA */
-    updateAndSimulateDepartedEvs(
-      currentTime,
-      departedEvsFromSimona,
-      tripProbabilities,
-      thresholdChargingHubDistance,
-      round15
-    )
-
     /* Get time until next event for one of the EVs and return corresponding tick to SIMONA */
-    val timeUntilNextEvent =
-      getTimeUntilNextEvent(electricVehicles, currentTime)
+    val timeUntilNextEvent = Seq(
+      getTimeUntilNextArrival(electricVehicles, currentTime),
+      getTimeUntilNextDeparture(electricVehicles, currentTime)
+    ).flatten.minOption
+
+    timeUntilNextEvent.foreach { nextEvent =>
+      logger.debug(
+        s"Next event in ${(nextEvent / 60 / 60).toInt} hours " +
+          s"and ${(nextEvent / 60) % 60} minutes."
+      )
+    }
 
     /* Save occupancy of charging stations for csv output */
     chargingStations.foreach(cs =>
@@ -138,16 +135,16 @@ final class MobilitySimulator(
     * @param maxDistance
     *   Maximum permissible distance between a POI and a charging station
     * @return
-    *   Returned EV models from SIMONA as set of Electric Vehicles
+    *   The updated charging station occupancy
     */
   private def exchangeEvMovementsWithSimona(
       currentTime: ZonedDateTime,
       availableChargingPoints: Map[UUID, Int],
       currentPricesAtChargingStations: Map[UUID, Double],
       maxDistance: Length
-  ): (Seq[ElectricVehicle], Map[UUID, Seq[ElectricVehicle]]) = {
+  ): Map[UUID, Seq[ElectricVehicle]] = {
 
-    /* Determine parking and departing evs in this tick */
+    // Determine parking and departing evs in this tick
     val (parkingEvs, departingEvs) =
       defineMovements(electricVehicles, currentTime)
 
@@ -170,7 +167,7 @@ final class MobilitySimulator(
       }
       .toSeq
 
-    /* Add EVs that start parking to movements and assign to Evcs UUID */
+    // Add EVs that start parking now to movements and assign to Evcs UUID
     val arrivals = handleParkingEvs(
       parkingEvs,
       currentPricesAtChargingStations,
@@ -178,12 +175,24 @@ final class MobilitySimulator(
       maxDistance
     )
 
+    // Update and simulate all EVs that were returned from SIMONA
+    updateAndSimulateDepartedEvs(
+      currentTime,
+      departedEvs,
+      tripProbabilities,
+      thresholdChargingHubDistance,
+      round15
+    )
+
+    // Finally, send arrivals to SIMONA
+    val nextArrival = getTimeUntilNextArrival(electricVehicles, currentTime)
     evData.provideArrivingEvs(
-      EvMovement.buildMovementsMap(arrivals)
+      EvMovement.buildMovementsMap(arrivals),
+      nextArrival.map(long2Long).toJava
     )
 
     // compile map from evcs to their parked evs
-    val evcsToParkedEvs = electricVehicles
+    electricVehicles
       .flatMap { ev =>
         ev.chosenChargingStation.map(ev -> _)
       }
@@ -192,8 +201,6 @@ final class MobilitySimulator(
       } { case (ev, _) =>
         ev
       }
-
-    (departedEvs, evcsToParkedEvs)
   }
 
   /** Determine the set of cars, that start to park and that depart in this time
@@ -501,53 +508,63 @@ final class MobilitySimulator(
 
   }
 
-  /** Return the time in seconds until the earliest next event for one of the
-    * EVs. The events are either departure or parking start.
+  /** Return the time in seconds until the earliest next departure for one of
+    * the EVs.
     *
     * @param evs
     *   Collection of all EVs
     * @param currentTime
     *   Current time to create a start value for foldLeft
     * @return
-    *   Time until earliest next event for one of the EVs in the set of all EVs
+    *   Time until earliest next departure for one of the EVs in the set of all
+    *   EVs
     */
-  private def getTimeUntilNextEvent(
+  private def getTimeUntilNextDeparture(
       evs: Seq[ElectricVehicle],
       currentTime: ZonedDateTime
-  ): Long = {
-
-    val nextEvent: ZonedDateTime = evs.foldLeft(currentTime.plusYears(1))(
-      (earliestEvent: ZonedDateTime, ev: ElectricVehicle) => {
-        val earlierEvent = if (ev.parkingTimeStart.isAfter(currentTime)) {
-          Seq(
-            earliestEvent,
-            ev.parkingTimeStart,
-            ev.departureTime
-          ).minOption.getOrElse(
-            throw new IllegalArgumentException(
-              "Unable to determine the earlier time"
-            )
-          )
-        } else {
-          Seq(earliestEvent, ev.departureTime).minOption.getOrElse(
-            throw new IllegalArgumentException(
-              "Unable to determine the earlier time"
-            )
-          )
-        }
-        earlierEvent
+  ): Option[Long] =
+    evs
+      .foldLeft[Option[ZonedDateTime]](None)(
+        (earliestEvent, ev: ElectricVehicle) =>
+          if (!ev.parkingTimeStart.isAfter(currentTime))
+            Seq(earliestEvent, Some(ev.departureTime)).flatten.minOption
+          else
+            earliestEvent
+      )
+      .map { nextEvent =>
+        currentTime.until(nextEvent, ChronoUnit.SECONDS)
       }
-    )
 
-    val timeUntilNextEvent: Long =
-      currentTime.until(nextEvent, ChronoUnit.SECONDS)
-    logger.debug(
-      s"Next event in ${(timeUntilNextEvent / 60 / 60).toInt} hours " +
-        s"and ${(timeUntilNextEvent / 60) % 60} minutes."
-    )
-
-    timeUntilNextEvent
-  }
+  /** Return the time in seconds until the earliest next arrival for one of the
+    * EVs.
+    *
+    * @param evs
+    *   Collection of all EVs
+    * @param currentTime
+    *   Current time to create a start value for foldLeft
+    * @return
+    *   Time until earliest next arrival for one of the EVs in the set of all
+    *   EVs
+    */
+  private def getTimeUntilNextArrival(
+      evs: Seq[ElectricVehicle],
+      currentTime: ZonedDateTime
+  ): Option[Long] =
+    evs
+      .foldLeft[Option[ZonedDateTime]](None)(
+        (earliestEvent, ev: ElectricVehicle) =>
+          if (ev.parkingTimeStart.isAfter(currentTime)) {
+            Seq(
+              earliestEvent,
+              Some(ev.parkingTimeStart),
+              Some(ev.departureTime)
+            ).flatten.minOption
+          } else
+            earliestEvent
+      )
+      .map { nextEvent =>
+        currentTime.until(nextEvent, ChronoUnit.SECONDS)
+      }
 
   private def updateElectricVehicles(movements: Seq[EvMovement]): Unit = {
     val movementMap = movements.map { case EvMovement(_, ev) =>